moved forms to separate files for easier testing
and added LoginForm tests
